% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lake_id_xref}
\alias{lake_id_xref}
\title{Lake identification cross-reference table}
\format{
\subsection{lake_id_xref}{

A data.frame with 489,557 rows and 8 columns
\describe{
\item{nhdhr.id}{Permanent identifier from NHDHR Resolution 2}
\item{nhd.comid}{COMID from NHDHR}
\item{winslow.id}{Permanent identifier from original NHD dataset}
\item{lagos.id}{Identifier from LAke multi-scaledGeOSpatial and temporal database}
\item{mglp.id}{Identifier from the Midwest Glacial Lakes Partnership}
\item{local.id}{The local identifier (usually state agency)}
\item{state}{The state that the local identifier corresponds to}
\item{agency}{The agency from which the local.id was taken}
}
}
}
\usage{
lake_id_xref
}
\description{
This table contains ID values for various different datasets that reference
these lakes. There are local (i.e. state ID fields), National Hydrography
Dataset (NHDHR), and NHD COMID. There is also a field for Winslow ID, which
relates to the original ID in the NHD dataset (i.e. pre-NHDHR)
}
\keyword{datasets}
