% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mwlaxeref-package.R
\docType{package}
\name{mwlaxeref-package}
\alias{mwlaxeref}
\alias{mwlaxeref-package}
\title{mwlaxeref: Cross-References Lake Identifiers Between Different Data Sets}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Handy helper package for cross-referencing lake identifiers among different data sets in the Midwestern United States. There are multiple different state, regional, and federal agencies that have different identifiers on lakes. This package helps you to go between them.
}
\author{
\strong{Maintainer}: Paul Frater \email{paul.frater@wisconsin.gov} (\href{https://orcid.org/0000-0002-7237-6563}{ORCID})

}
\keyword{internal}
