% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wis_lakes}
\alias{wis_lakes}
\title{Example of lake ID data from Wisconsin}
\format{
\subsection{wis_lakes}{

A data.frame with 10 rows and 4 columns
\describe{
\item{state}{The state abbreviation (i.e. "wi")}
\item{county}{The name of the county in which the lake is found}
\item{lake.id}{A unique identifier for each lake}
\item{lake.name}{The name of the lake}
}
}
}
\usage{
wis_lakes
}
\description{
This data.frame contains the names and lake IDs for 10 lakes in
Wisconsin. These are used in examples throughout the package.
}
\keyword{datasets}
